# apps/api/v2/swagger/documentation.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v2/bookings": {
      "get": {
        "operationId": "BookingsController_getBookings",
        "summary": "Get user bookings",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "filters[status]",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "upcoming",
                "recurring",
                "past",
                "cancelled",
                "unconfirmed"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBookingsOutput"
                }
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ]
      },
      "post": {
        "operationId": "BookingsController_createBooking",
        "summary": "Book an event",
        "parameters": [
          {
            "name": "x-cal-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/v2/bookings/{bookingUid}": {
      "get": {
        "operationId": "BookingsController_getBooking",
        "summary": "Get booking details by UID",
        "parameters": [
          {
            "name": "bookingUid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBookingOutput"
                }
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/v2/bookings/{bookingUid}/reschedule": {
      "get": {
        "operationId": "BookingsController_getBookingForReschedule",
        "summary": "Get booking details by UID to reschedule",
        "parameters": [
          {
            "name": "bookingUid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/v2/bookings/{bookingId}/cancel": {
      "post": {
        "operationId": "BookingsController_cancelBooking",
        "summary": "Cancel an event booking",
        "parameters": [
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-cal-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelBookingInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/v2/bookings/{bookingUid}/mark-no-show": {
      "post": {
        "operationId": "BookingsController_markNoShow",
        "summary": "Mark host/attendee as no-show",
        "parameters": [
          {
            "name": "bookingUid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkNoShowInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkNoShowOutput"
                }
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/v2/bookings/recurring": {
      "post": {
        "operationId": "BookingsController_createRecurringBooking",
        "summary": "Book a recurring event",
        "parameters": [
          {
            "name": "x-cal-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Create Recurring Booking Input",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateRecurringBookingInput"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/v2/bookings/instant": {
      "post": {
        "operationId": "BookingsController_createInstantBooking",
        "summary": "Create an instant booking",
        "parameters": [
          {
            "name": "x-cal-client-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
...
  },
  "info": {
    "title": "OneHash Cal v2 API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Location": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "weekStart": {
            "type": "string"
          },
          "organizationId": {
            "type": "number"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "bookerUrl": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "name",
          "weekStart",
          "profile",
          "bookerUrl"
        ]
      },
...
      "Attendee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "locale": {
            "type": "string",
            "nullable": true
          },
          "bookingId": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "timeZone",
          "locale",
          "bookingId"
        ]
      },
...
      "GetBookingsDataEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "userPrimaryEmail": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "customInputs": {
            "type": "object"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "attendees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attendee"
            }
          },
          "metadata": {
            "type": "object"
          },
          "uid": {
            "type": "string"
          },
          "recurringEventId": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "eventType": {
            "$ref": "#/components/schemas/EventType"
          },
          "status": {
            "type": "object"
          },
          "paid": {
            "type": "boolean"
          },
          "payment": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "references": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reference"
            }
          },
          "isRecorded": {
            "type": "boolean"
          },
          "seatsReferences": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "user": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "rescheduled": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "customInputs",
          "startTime",
          "endTime",
          "attendees",
          "metadata",
          "uid",
          "recurringEventId",
          "location",
          "eventType",
          "status",
          "paid",
          "payment",
          "references",
          "isRecorded",
          "seatsReferences",
          "user"
        ]
      },
      "GetBookingsData": {
        "type": "object",
        "properties": {
          "bookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetBookingsDataEntry"
            }
          },
          "recurringInfo": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "nextCursor": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "bookings",
          "recurringInfo",
          "nextCursor"
        ]
      },
      "GetBookingsOutput": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/GetBookingsData"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
      "GetBookingData": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "id": {
            "type": "number"
          },
          "uid": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "customInputs": {
            "type": "object"
          },
          "smsReminderNumber": {
            "type": "string",
            "nullable": true
          },
          "recurringEventId": {
            "type": "string",
            "nullable": true
          },
          "startTime": {
            "format": "date-time",
            "type": "string"
          },
          "endTime": {
            "format": "date-time",
            "type": "string"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          },
          "responses": {
            "type": "object"
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true
          },
          "userPrimaryEmail": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "attendees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attendee"
            }
          },
          "eventTypeId": {
            "type": "number",
            "nullable": true
          },
          "eventType": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/EventType"
              }
            ]
          }
        },
        "required": [
          "title",
          "id",
          "uid",
          "description",
          "customInputs",
          "smsReminderNumber",
          "recurringEventId",
          "startTime",
          "endTime",
          "location",
          "status",
          "metadata",
          "cancellationReason",
          "responses",
          "rejectionReason",
          "userPrimaryEmail",
          "user",
          "attendees",
          "eventTypeId",
          "eventType"
        ]
      },
      "GetBookingOutput": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/GetBookingData"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
      "Response": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "guests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "guests"
        ]
      },
      "CreateBookingInput": {
        "type": "object",
        "properties": {
          "end": {
            "type": "string"
          },
          "start": {
            "type": "string"
          },
          "eventTypeId": {
            "type": "number"
          },
          "eventTypeSlug": {
            "type": "string"
          },
          "rescheduleUid": {
            "type": "string"
          },
          "recurringEventId": {
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "user": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "language": {
            "type": "string"
          },
          "bookingUid": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "hasHashedBookingLink": {
            "type": "boolean"
          },
          "hashedLink": {
            "type": "string",
            "nullable": true
          },
          "seatReferenceUid": {
            "type": "string"
          },
          "responses": {
            "$ref": "#/components/schemas/Response"
          },
          "orgSlug": {
            "type": "string"
          },
          "locationUrl": {
            "type": "string"
          }
        },
        "required": [
          "start",
          "eventTypeId",
          "timeZone",
          "language",
          "metadata",
          "hashedLink",
          "responses"
        ]
      },
      "CancelBookingInput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "uid": {
            "type": "string"
          },
          "allRemainingBookings": {
            "type": "boolean"
          },
          "cancellationReason": {
            "type": "string"
          },
          "seatReferenceUid": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "uid",
          "allRemainingBookings",
          "cancellationReason",
          "seatReferenceUid"
        ]
      },
      "MarkNoShowInput": {
        "type": "object",
        "properties": {
          "noShowHost": {
            "type": "boolean"
          },
          "attendees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attendee"
            }
          }
        }
      },
      "HandleMarkNoShowData": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "noShowHost": {
            "type": "boolean"
          },
          "attendees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attendee"
            }
          }
        },
        "required": [
          "message"
        ]
      },
      "MarkNoShowOutput": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success",
            "enum": [
              "success",
              "error"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/HandleMarkNoShowData"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
      "CreateRecurringBookingInput": {
        "type": "object",
        "properties": {
          "end": {
            "type": "string"
          },
          "start": {
            "type": "string"
          },
          "eventTypeId": {
            "type": "number"
          },
          "eventTypeSlug": {
            "type": "string"
          },
          "rescheduleUid": {
            "type": "string"
          },
          "recurringEventId": {
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "user": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "language": {
            "type": "string"
          },
          "bookingUid": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "hasHashedBookingLink": {
            "type": "boolean"
          },
          "hashedLink": {
            "type": "string",
            "nullable": true
          },
          "seatReferenceUid": {
            "type": "string"
          },
          "responses": {
            "$ref": "#/components/schemas/Response"
          },
          "orgSlug": {
            "type": "string"
          },
          "locationUrl": {
            "type": "string"
          },
          "noEmail": {
            "type": "boolean"
          },
          "recurringCount": {
            "type": "number"
          },
          "appsStatus": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "allRecurringDates": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "currentRecurringIndex": {
            "type": "number"
          }
        },
        "required": [
          "start",
          "eventTypeId",
          "timeZone",
          "language",
          "metadata",
          "hashedLink",
          "responses"
        ]
      },
...
    }
  }
}